Check strconv.ParseUint error.
`strconv.ParseUint` returns 0 and an error if argument is invalid. If we
don't check error, the following `chgrd` will succeed to change the
group of file to 0, and reports no error if the process is
running as root user. This is not expected.

(cherry picked from commit ed558c72c00e10822767f98f62af3eeae2482e4b)