Expose `shouldProvision` to provisioners via new controller interface
Added new `controller.Qualifier` interface which is implemented by
any controller that has a `ShouldProvision` method. When implemented,
the provisioner's `ShouldProvision` method is called by lib controller's
`shouldProvision`, with a return value of false from the former
resulting in the same from the latter.

This allows provisioners the opportunity to ignore individual claims
prior to attaining leadership for the claim, avoiding uneccessary
election cycles when multiple provisioners are serving the same storage
class.