update deployment instructions and version updates
Minor edits to the README and deployment instructions
to indicate that the migration tasks are still underway.

Till the new image repo is setup and images are pushed,
the `make container` needs to be customed to generate image
to a custom location and should be used with the deployment.

Also updated the minimum K8s version required to 1.14 to allow
updating the spec version for PSP and Claim.

Signed-off-by: kmova <kiran.mova@mayadata.io>